Added credit_cap table to db
