+++ src/api/db/admin.py
@@ -40,25 +40,3 @@ class Admin:
 			return(True, None)
 		else:
 			return (False, error)
-
-	def set_credit_cap(self, credit_cap, warning_message):
-		try:
-			delete = "DELETE FROM credit_cap;"
-			response, error = self.db_conn.execute(delete, None, isSELECT=False)
-		except Exception as e:
-			return (False, e)
-		if response == None:
-			return (False, error)
-
-		try:
-			cmd = """
-				INSERT INTO credit_cap(credit_cap, credit_cap_warning)
-				VALUES(%s, %s)
-			"""
-			response, error = self.db_conn.execute(cmd, [credit_cap, warning_message], False)
-		except Exception as e:
-			return (False, e)
-		
-		if response != None:
-			return (True, None)
-		return (False, error)
